Improved functionality with generated asteroids
