+++ 3d Space Rocks 2020.py
@@ -1,22 +1,4 @@
 from direct.showbase.ShowBase import ShowBase
-from panda3d.core import *
-from direct.interval.IntervalGlobal import * 
-from direct.gui.OnscreenText import OnscreenText
-from direct.gui.DirectGui import *
-from direct.showbase.DirectObject import DirectObject
-from direct.actor.Actor import Actor
-from direct.task import Task 
-from direct.task.Task import Task
-from pprint import pprint
-import math
-import sys
-import random
-import collections
-from opensimplex import OpenSimplex as opens
-import time
-
-
-from direct.showbase.ShowBase import ShowBase
 from panda3d.core import Geom, GeomNode, GeomVertexFormat, \
     GeomVertexData, GeomTriangles, GeomVertexWriter, GeomVertexReader
 from panda3d.core import NodePath
@@ -30,43 +12,60 @@ import math
 import collections
 from opensimplex import OpenSimplex as opens
 import time
+from direct.gui.OnscreenText import OnscreenText
+from direct.interval.IntervalGlobal import *
 
-print("Starting 3d Asteroid Simulator 2020")
-asteroid_spawn_distance = 9900000 # How close the asteroids spawn and how far away they will fly to
+asteroid_spawn_distance     = 990000 # How close the asteroids spawn and how far away they will fly to
 asteroid_min_spawn_distance = 5000
-asteroid_speed = 90 # Time it takes an asteroid to fly from one side to another in seconds
-asteroid_speed_percent = 30 # the pecentage that asteroid_speed can be modified by randomly for each asteroid. 
-asteroid_future_distance = 2000000 # How far the asteroid will travel
-asteroid_total = []
-missle_total = []
-asteroid_max = 200 # The maximum number of asteroids ***** MUST BE AN EVEN NUMBER  and make sur eto modify loop_test_number global variable*****
-spaceship_speed_const = 40
-spaceship_speed_x = 0
-spaceship_speed_y = 0
-spaceship_speed_z = 0
-colors = {"orange": (.9,.6,.05,1), "gray": (.1,.1,.1,1), "black": (0,0,0,1), "white": (1,1,1,1), "white-transparent": (1,1,1,0.4), "red": (1,0,0,1), "red-transparent": (1,0,0,0.4), "yellow-tinge": (1,1,0.8,1), "yellow-tinge-transparent": (1,1,0.8,0.4)}
-loop_test = 0 # Used to limit the number of asteroids that have their distance tested each frame
-loop_test_number = 0 # The number of asteroids to test their distance each frame ***** ASTEROID_MAX MUST BE EVENLY DIVISIBLE BY THIS NUMBER - WILL FIX THIS LATER *****
-last_frame_mpos_x = 0
-last_frame_mpos_y = 0
-asteroid_test_distance = 990000 #The test distance. If asteroid greater than asteroid_test_distancem then it will be moved closer
-score = 0  # Initialize score
+asteroid_future_distance    = 2000000 # How far the asteroid will travel
+asteroid_total              = []
+missle_total                = []
+asteroid_max                = 1000 # The maximum number of asteroids ***** MUST BE AN EVEN NUMBER  and make sur eto modify loop_test_number global variable*****
+spaceship_speed_const       = 50
+spaceship_speed_x           = 0
+spaceship_speed_y           = 0
+spaceship_speed_z           = 0
+colors                      = {"orange": (.9,.6,.05,1), "gray": (.1,.1,.1,1), "black": (0,0,0,1), "white": (1,1,1,1), "white-transparent": (1,1,1,0.4), "red": (1,0,0,1), "red-transparent": (1,0,0,0.4), "yellow-tinge": (1,1,0.8,1), "yellow-tinge-transparent": (1,1,0.8,0.4)}
+loop_test                   = 0 # Used to limit the number of asteroids that have their distance tested each frame
+loop_test_number            = 40 # The number of asteroids to test their distance each frame ***** ASTEROID_MAX MUST BE EVENLY DIVISIBLE BY THIS NUMBER - WILL FIX THIS LATER *****
+last_frame_mpos_x           = 0
+last_frame_mpos_y           = 0
+asteroid_test_distance      = 999000 #The test distance. If asteroid greater than asteroid_test_distancem then it will be moved closer
+score                       = 0  # Initialize score
+FullSceeen                  = False
+Frames                      = True
 
-class begin(ShowBase):
+class Begin(ShowBase):
     def __init__(self):
+        # Basics
         ShowBase.__init__(self)
         #Setup the window
         base.disableMouse()
-        base.setFrameRateMeter(True)
-        base.camLens.setFar(asteroid_spawn_distance * 2)
+        if Frames:
+            base.setFrameRateMeter(True)
+        base.camLens.setFar(asteroid_spawn_distance * 100)
         render.setAntialias(AntialiasAttrib.MAuto)
-        props = WindowProperties()
-        props.setCursorHidden(True)
-        self.win.requestProperties(props)
+        wp = WindowProperties()
+        wp.setSize(800,600)
+        wp.setCursorHidden(True)
+        self.win.requestProperties(wp)
         self.setBackgroundColor(colors.get("black"))
+        # Lights
+        ambientLight = AmbientLight("ambientLight")
+        ambientLight.setColor((0.5, 0.5, 0.5, 1))
+        directionalLight = DirectionalLight("directionalLight")
+        directionalLight.setDirection(LVector3(0, 45, -45))
+        directionalLight.setColor((1, 1, 1, 1))
+        directionalLight.setShadowCaster(True)
+        render.setLight(render.attachNewNode(directionalLight))
+        render.setLight(render.attachNewNode(ambientLight))
+        #fog
+        self.fog = Fog('distanceFog')
+        self.fog.setColor(0, 0, 0)
+        self.fog.setExpDensity(.000003)
+        render.setFog(self.fog)
 
-
-        begin.keyMap = {
+        Begin.keyMap = {
             "forward": False, "strafe-left": False, "backward": False, "strafe-right": False, "strafe-up": False, "strafe-down": False, "roll-left": False, "roll-right": False} #True if coresponding key is currently held down.
         #Basic camera movement on the xyz coordinate plane
         self.accept("escape", sys.exit)
@@ -86,60 +85,28 @@ class begin(ShowBase):
         self.accept("q-up", self.setKey, ["roll-left", False])
         self.accept("e", self.setKey, ["roll-right", True])
         self.accept("e-up", self.setKey, ["roll-right", False])
-        # Development keys
-        self.accept('k', self.createAsteroids) # K for Kreate - Creates asteroids
-        self.accept('n', self.nuke) # N for Nuke - Removes all asteroids
-        self.accept('p', self.nuke_missles) #P for pop missles
         self.accept('mouse1', self.shoot) # Shoots the projectile
+        self.accept('f11', self.fullscreenToggle)
+        self.accept('f12', self.framesToggle)
+        # Development keys
         self.accept('0', self.stop_moving) # Stop moving
 
-        ambientLight = AmbientLight("ambientLight")
-        ambientLight.setColor((.02, 0.02, 0.02, 1))
-        directionalLight = DirectionalLight("directionalLight")
-        directionalLight.setDirection(LVector3(0, 45, -45))
-        render.setLight(render.attachNewNode(directionalLight))
-        render.setLight(render.attachNewNode(ambientLight))
-        # Create the World
-        #self.setupLights()
-        #self.createFog()
-        #self.startTasks()
-        #self.test_asteroid()
-        #self.createAsteroids()
-        asteroid = Asteroid()
-        print(render.findAllMatches("*"))
+        # Create the geometry
+        self.startTasks()
+        self.createAsteroids()
 
-    def setupLights(self):  # This function sets up some default lighting
-        ambientLight = AmbientLight("ambientLight")
-        ambientLight.setColor((.02, 0.02, 0.02, 1))
-        directionalLight = DirectionalLight("directionalLight")
-        directionalLight.setDirection(LVector3(0, 45, -45))
-        render.setLight(render.attachNewNode(directionalLight))
-        render.setLight(render.attachNewNode(ambientLight))
-
-    def createFog(self):
-        fog = Fog('distanceFog')
-        fog.setColor(0, 0, 0)
-        fog.setExpDensity(.000004)
-        render.setFog(fog)
+    def createAsteroids(self):
+        for i in range(0,asteroid_max):
+            print(int((i / asteroid_max)*100) , "% done")
+            asteroid_total.append(Asteroid(random.choice(["small","medium","large"])))
 
     def startTasks(self):
         #The tasks below are the functions run every frame so the game will work
-        taskMgr.add(begin.test_distance, "Test Distance")
-        taskMgr.add(begin.mouseTask, "Camera Look at Mouse")
-        taskMgr.add(begin.spaceship_movement, "Move the Player or 'spaceship' ")
-        taskMgr.add(begin.remove_old_missles, "Remove old missles")
-        taskMgr.add(begin.score, "Score")
-
-    def test_asteroid(self):
-        self.modelName = "asteroid" + str(random.randrange(1,25,1)) # Name must be at least 3 charachters long
-        self.obj = loader.loadModel("./Models/" + self.modelName + ".egg")
-        self.obj.reparentTo(render)
-        self.obj.setColor(colors.get("gray")) #(gray, gray, gray, 1)
-        self.obj.setScale((random.randrange(900, 9000, 1) ,random.randrange(900, 9000, 1) ,random.randrange(900, 9000, 1)))
-
-    def createAsteroids(self, number=asteroid_max): # Can be used to create as many asteroids as wanted. Defalut is 500 for testing purposes.
-        for i in range(number):
-            asteroid_total.insert(0,Asteroid())
+        taskMgr.add(Begin.test_distance, "Test Distance")
+        taskMgr.add(Begin.mouseTask, "Camera Look at Mouse")
+        taskMgr.add(Begin.spaceship_movement, "Move the Player or 'spaceship' ")
+        taskMgr.add(Begin.remove_old_missles, "Remove old missles")
+        taskMgr.add(Begin.score, "Score")
 
     ##### // Key Press Functions \\ #####
     def spaceship_movement(self):
@@ -152,52 +119,52 @@ class begin(ShowBase):
         base.camera.setPos(base.camera, spaceship_speed_x, # Spaceship X change per frame
                                         spaceship_speed_y, # Spaceship Y change per frame
                                         spaceship_speed_z) #   "       Z   "     "    "
-        if begin.keyMap["forward"]:
+        if Begin.keyMap["forward"]:
             if spaceship_speed_y < 0:
                 spaceship_speed_y += 40 * math.log(-spaceship_speed_y)
             elif spaceship_speed_y < 1:
                 spaceship_speed_y = 1.1
             else:
-                spaceship_speed_y = 1000 * math.log(spaceship_speed_y)
-        if begin.keyMap["backward"]:
+                spaceship_speed_y = 800 * math.log(spaceship_speed_y)
+        if Begin.keyMap["backward"]:
             if spaceship_speed_y > 0:
                 spaceship_speed_y -= 40 * math.log(spaceship_speed_y)
             elif spaceship_speed_y == 0:
                 spaceship_speed_y = -1.1
             else:
                 spaceship_speed_y = -(100 * math.log(-spaceship_speed_y))
-        if begin.keyMap["strafe-left"]:
+        if Begin.keyMap["strafe-left"]:
             if spaceship_speed_x > 0:
                 spaceship_speed_x -= 40 * math.log(spaceship_speed_x)
             elif spaceship_speed_x == 0:
                 spaceship_speed_x = -1.1
             else:
                 spaceship_speed_x = -(100 * math.log(-spaceship_speed_x))
-        if begin.keyMap["strafe-right"]:
+        if Begin.keyMap["strafe-right"]:
             if spaceship_speed_x < 0:
                 spaceship_speed_x += 40 * math.log(-spaceship_speed_x)
             elif spaceship_speed_x < 1:
                 spaceship_speed_x = 1.1
             else:
                 spaceship_speed_x = 100 * math.log(spaceship_speed_x)
-        if begin.keyMap["strafe-up"]:
+        if Begin.keyMap["strafe-up"]:
             if spaceship_speed_z < 0:
                 spaceship_speed_z += 40 * math.log(-spaceship_speed_z)
             elif spaceship_speed_z < 1:
                 spaceship_speed_z = 1.1
             else:
                 spaceship_speed_z = 100 * math.log(spaceship_speed_z)
-        if begin.keyMap["strafe-down"]:
+        if Begin.keyMap["strafe-down"]:
             if spaceship_speed_z > 0:
                 spaceship_speed_z -= 40 * math.log(spaceship_speed_z)
             elif spaceship_speed_z == 0:
                 spaceship_speed_z = -1.1
             else:
                 spaceship_speed_z = -(100 * math.log(-spaceship_speed_z))
-        if begin.keyMap["roll-left"]:
+        if Begin.keyMap["roll-left"]:
             camera_r = base.camera.getR()
             base.camera.setR(camera_r - 1)
-        if begin.keyMap["roll-right"]:
+        if Begin.keyMap["roll-right"]:
             camera_r = base.camera.getR()
             base.camera.setR((camera_r + 1))
         return Task.cont
@@ -227,7 +194,6 @@ class begin(ShowBase):
             distance = math.sqrt((asteroid_xyz[0] - camera_xyz[0])**2 + (asteroid_xyz[1] - camera_xyz[1])**2 + (asteroid_xyz[2] - camera_xyz[2])**2) # Distance formula
             if distance > asteroid_test_distance: 
                 future_pos = asteroid.get_sphere_points(asteroid_spawn_distance, base.camera)
-                #print("Asteroid #", asteroid.obj.name , " is too faw away at " , distance , "Moving to " , future_pos)
                 asteroid.asteroid_lerp.finish()
                 asteroid.asteroid_path(future_pos) #move to sphere relative to camera
 
@@ -265,6 +231,7 @@ class begin(ShowBase):
         for missle in missle_total:
             if missle.ttl <= 0:
                 render.clearLight(missle.plnp)
+                missle.core.detachNode()
             else:
                 missle.ttl -= dt
         return Task.cont
@@ -302,22 +269,60 @@ class begin(ShowBase):
     def setKey(self, key, value):
         self.keyMap[key] = value
 
+    def fullscreenToggle(self):
+        global FullSceeen
+        global Frames
+        wp = WindowProperties()
+        if (not(FullSceeen)):
+            wp.setFullscreen(True)
+            wp.setCursorHidden(True)
+            wp.setSize(3840,2160)
+            base.openMainWindow()
+            base.win.requestProperties(wp)
+            base.graphicsEngine.openWindows()
+            FullSceeen = True
+        else:
+            wp.setFullscreen(False)
+            wp.setCursorHidden(True)
+            wp.setSize(800,600)
+            base.openMainWindow()
+            base.win.requestProperties(wp)
+            base.graphicsEngine.openWindows()
+            FullSceeen = False
+        #Resizing clears frames. Reset that
+        base.setFrameRateMeter(Frames)
+
+    def framesToggle(self):
+        global Frames
+        if(Frames):
+            base.setFrameRateMeter(False)
+            Frames = False
+        else:
+            base.setFrameRateMeter(True)
+            Frames = True
 
 class Asteroid(object):
-    def __init__(self, step=36, radius=100000):
+
+    def __init__(self, size, step=36):
+        types = {"large":  {"radius": 10000, "darkest_gray": 0.3, "lightest_gray": 0.7, "speed": 9000, "speed_percent": 5},
+                 "medium": {"radius": 5000,  "darkest_gray": 0.4, "lightest_gray": 0.7, "speed": 900,  "speed_percent": 10},
+                 "small":  {"radius": 1000,   "darkest_gray": 0.4, "lightest_gray": 0.7, "speed": 90,   "speed_percent": 20}}
+        #small should be 1000 -> 1500
         #Step for creating the sphere, lower the number greater the detail of the shpere
-        self.step =  step
-        self.radius = radius
-        self.map = self.create_map(self.radius)
-        self.asteroid_min = 10000
-        self.asteroid_max = 15000
-        print("map created")
-        geom = self.create_geom(self.radius)
-        print("geom created")
-        np = NodePath(geom)
-        print("nodepath created")
-        #np.reparent_to(render)
-        np.reparentTo(render)
+        self.radius         = types[size]["radius"]
+        self.darkest_gray   = types[size]["darkest_gray"]
+        self.lightest_gray  = types[size]["lightest_gray"]
+        self.speed          = types[size]["speed"]
+        self.speed_percent  = types[size]["speed_percent"]
+        self.step           = step
+        self.asteroid_min   = self.radius
+        self.asteroid_max   = self.radius + (self.radius / 2)
+        self.map            = self.create_map(self.radius)
+        geom                = self.create_geom(self.radius)
+        self.np             = NodePath(geom)
+        spawn_distance      = self.translate((random.random() ** 0.5),0,1,asteroid_min_spawn_distance, asteroid_spawn_distance)
+        self.asteroid_path(self.get_sphere_points(spawn_distance))
+        self.np.reparent_to(render)
 
 
     def create_geom(self, sidelength):
@@ -334,6 +339,7 @@ class Asteroid(object):
         for key , value in self.map.items():
             v_x, v_y, v_z, asteroid_color = value
             n_x, n_y, n_z = 1,1,1
+            #c_r, c_g, c_b, c_a = asteroid_color, asteroid_color, asteroid_color, 1
             c_r, c_g, c_b, c_a = asteroid_color, asteroid_color, asteroid_color, 1
             vertex.addData3f(v_x, v_y, v_z)
             normal.addData3f(n_x, n_y, n_z)
@@ -346,7 +352,6 @@ class Asteroid(object):
             tris.addVertices(vert + 1, 0, vert)
             tris.closePrimitive()
             geom.addPrimitive(tris)
-        print("top created")
         #middle of shpere
         for row in range(1, int(180 / self.step) - 1):
             for vert_iir in range(0, verts_per_row - 1): # vert_iir = vertex index in row, not vertex number
@@ -362,7 +367,6 @@ class Asteroid(object):
                 tris.add_vertices(vert_number + 1, vert_up_row + 1, vert_number)
                 tris.close_primitive()
                 geom.addPrimitive(tris)
-        print("middle created")
         #bottom of sphere
         last_vert = len(self.map) - 1
         for vert in range(last_vert - verts_per_row, last_vert):
@@ -387,21 +391,21 @@ class Asteroid(object):
 
     def simplex_radius(self, radius, seed, xoff, yoff, zoff):
         noise = opens(seed=seed)
-        return self.translate(noise.noise3d(xoff,yoff,zoff),0,1,10,15)
+        return self.translate(noise.noise3d(xoff,yoff,zoff),0,1,self.asteroid_min, self.asteroid_max)
         #return radius
 
     #map for most of the sphere
     def create_map(self, radius):
         map = collections.OrderedDict()
-        seed = int(time.time())
+        seed = int(time.time() * 10000000)
         for x in range(0, 181, self.step):
             #Top/Bottom of sphere need to be single points
             if x == 0:
                 point_radius = self.simplex_radius(radius, seed, 0, 0, 1) 
-                map[(0,0)] = (0,0,point_radius, (self.translate(point_radius, 10000, 15000, 0.3, 0.5)))
+                map[(0,0)] = (0,0,point_radius, (self.translate(point_radius, self.asteroid_min, self.asteroid_max, self.darkest_gray, self.lightest_gray)))
             elif x == 180:
                 point_radius = -(self.simplex_radius(radius, seed, 0, 0, -1)) 
-                map[(180,0)] = (0,0,point_radius, (self.translate(point_radius, -10000, -15000, 0.3, 0.5)))
+                map[(180,0)] = (0,0,point_radius, (self.translate(point_radius, -self.asteroid_min, -self.asteroid_max, self.darkest_gray, self.lightest_gray)))
             #The rest of the sphere
             else:
                 for y in range(0, 361, self.step):
@@ -414,10 +418,30 @@ class Asteroid(object):
                     v_x = point_radius * xoff
                     v_y = point_radius * yoff
                     v_z = point_radius * zoff
-                    map[(x ,y)] = (v_x, v_y, v_z, (self.translate(point_radius, 10000, 15000, 0.3, 0.5)))
-                    #print(map[(x,y)])
+                    map[(x ,y)] = (v_x, v_y, v_z, (self.translate(point_radius, self.asteroid_min, self.asteroid_max, self.darkest_gray, self.lightest_gray)))
         return map 
 
+    def asteroid_path(self, start_point): #takes starting location (start_point must be a LPoint3)
+        # Create and run the asteroid animation
+        self.asteroid_lerp = LerpPosInterval(self.np, # Object being manipulated. The asteroid in this case.
+                                            self.speed * random.randrange(1, self.speed_percent, 1), # How fast the asteroid will move in seconds
+                                            self.get_sphere_points(asteroid_future_distance, base.camera), # future location at end of lerp
+                                            start_point) # The start position of the asteroid
+        self.asteroid_lerp.start()
+
+    def get_sphere_points(self, radius, relative_to=False): #returns a LPoint3 in sphere. relative_to will return global LPoint3 realtive to given object.
+        phi = random.uniform(0,2*math.pi)
+        theta = random.uniform(0,2*math.pi)
+        if relative_to:
+            rel_xyz = relative_to.getPos()
+            x = radius * math.cos(phi) * math.sin(theta) + rel_xyz[0]
+            y = radius * math.sin(phi) * math.sin(theta) + rel_xyz[1]
+            z = radius * math.cos(theta) + rel_xyz[2]
+        else:
+            x = radius * math.cos(phi) * math.sin(theta)
+            y = radius * math.sin(phi) * math.sin(theta)
+            z = radius * math.cos(theta)
+        return LPoint3(x,y,z)
 
 class Location(object): # Create child of asteroid to find the future position the asteroid will fly to
     def __init__(self, asteroid_parent, distance=asteroid_future_distance):
@@ -426,11 +450,12 @@ class Location(object): # Create child of asteroid to find the future position t
 
 class Missle(object):
     def __init__(self):
+        print("missle created")
         camera_hpr = base.camera.getHpr()
         self.core = loader.loadModel("./Models/sphere.egg")
         self.core.setPos(base.camera, (0,0,0))
         self.core.setHpr(camera_hpr)
-        self.ttl = 0.5 # Time to live in seconds
+        self.ttl = 1 # Time to live in seconds
         self.core.setScale(600,600,600)
         self.glow = loader.loadModel("./Models/sphere.egg")
         self.core.setTransparency(TransparencyAttrib.MAlpha)
@@ -445,7 +470,8 @@ class Missle(object):
         #Create the light so the missle glows
         plight = PointLight('plight')
         plight.setColor(colors.get("white"))
-        plight.setAttenuation(LVector3(0, 0.000005 , 0))
+        plight.setAttenuation(LVector3(0, 0.00008 , 0))
+        plight.setMaxDistance(100)
         self.plnp = self.core.attachNewNode(plight) #point light node point
         render.setLight(self.plnp)
         """for asteroid in asteroid_total:
@@ -462,35 +488,5 @@ class Missle(object):
 
         self.core.reparentTo(render)
 
-bigBucks = begin()
-bigBucks.run()
-
-"""
-**********           Done           **********
-> Spawn asteroids in sphere
-> Continue to spawn asteroids around the ship as flown through space
-> Add missle
-> Add asteroid movement animation
-> Give missles glow, and remove old missles so more than 13 will glow
-> Add basic score UI
-
-********** Things that I need to do **********
-
-> Detect when missle hits asteroid, then add a point
-> Generate asteroids in rectangular, sphereical, and pill shape
-> create 3 different sizes of asteroids w/ different speed
-    > At begininning all should be large, and med/small will only show becasue of shooting.
-> Add rotation to asteroids
-> Break asteroids upon impact with missle
-> Asteroids should break upon impact w/ eachother, but not give points.
-> Add a fail state to the game.
-> Organize the code, I should do this now, but I am lazy
-> Upgrade point system to be relative to ship speed, and density/speed of asteroid (easy, medium, hard)
-> Add a delay in space ship turning to make the game better
-> Add hud that shows points, global direction of ship, and lag between mouse and direction of ship
-> Particle effects
-> Main Menu (Dificulty select, and start) it would be cool if it looked like the original asteroids.
-    > This should include a name on the window.
-> Add a fail menu
-> Add loading menu (should be quick, but looks good.)
-"""
+demo = Begin()
+demo.run()
